Documented the Games module. Changed the ID of the game to use UUID instead of a random string hence the deletion of StringUtils.java
Changed the SystemResponse class to be a record instead of a class.

Added GameResponseContentKeys enumeration to hold all the keys to avoid hardcoding them in the tests.

Refactored the FixedMovementPlayerTest to supply arguments using a provider in order to parametrize the test run. Also, improved the accuracy of the check of the randomness.